clone deps in pull and migrate (#707)
